fix(quicknode): fix fee estimate (#266)
* fix: fix unit conversion + use bignumber.js instead of big.js

The `btcToSats` function was a bit fragile. It was checking the input as a
string to make sure were not having more than 8 digits after the decimal
point.

However, the `.toString()` methods might be using scientific notation
when converting some numbers, which would be an error with the previous
check we made.

Instead, we now use `BigNumber` which also supports integers and we
use this to make sure BTC numbers are within range.

* fix(quicknode): fix fee estimation (kvB -> vB)

* chore: lint

* chore: update bitcoin core comments

Co-authored-by: Stanley Yuen <102275989+stanleyyconsensys@users.noreply.github.com>

* refactor: add big prefix when using BigNumber(s)

* refactor: btcSatsFactor > satsIn1Btc

* fix(unit): check if kvb can be safely converted to vb + round up

* test: add tests for satsKVBToVB

---------

Co-authored-by: Stanley Yuen <102275989+stanleyyconsensys@users.noreply.github.com>